Delete exception

+++ src/main/java/com/educandoWeb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoWeb.course.entities.User;
 import com.educandoWeb.course.repositories.UserRepository;
+import com.educandoWeb.course.services.exceptions.DataBaseException;
 import com.educandoWeb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,8 +32,17 @@ public class UserService {
 	}
 	
 	public void deleteById(Long Id) {
-		userRepository.deleteById(Id);
-	}
+		try {
+			if(userRepository.existsById(Id)) {
+				userRepository.deleteById(Id);
+			}else {
+				throw new ResourceNotFoundException(Id);
+			}
+	}catch(ResourceNotFoundException e) {
+		throw new ResourceNotFoundException(Id);
+	}catch(DataIntegrityViolationException e) {
+		throw new DataBaseException(e.getMessage());
+	}}
 	
 	public User update(Long id, User obj) {
 		User entity = userRepository.getReferenceById(id);
